+++ examples/cookie-sessions/index.js
@@ -4,17 +4,19 @@
  */
 
 var express = require('../../');
+var favicon = require('static-favicon');
+var cookie-parser = require('cookie-parser');
 
 var app = module.exports = express();
 
 // ignore GET /favicon.ico
-app.use(express.favicon());
+app.use(favicon());
 
 // pass a secret to cookieParser() for signed cookies
-app.use(express.cookieParser('manny is cool'));
+app.use(cookieParser('manny is cool'));
 
 // add req.session cookie support
-app.use(express.cookieSession());
+app.use(cookieSession());
 
 // do something with the session
 app.use(count);