+++ examples/cookies/app.js
@@ -4,7 +4,11 @@
  */
 
 var express = require('../../')
-  , app = module.exports = express();
+  , app = module.exports = express()
+  , favicon = require('static-favicon')
+  , logger = require('morgan')
+  , cookieParser = require('cookie-parser')
+  , bodyParser = require('body-parser')
 
 
 // add favicon() before logger() so
@@ -12,20 +16,20 @@ var express = require('../../')
 // logged, because this middleware
 // reponds to /favicon.ico and does not
 // call next()
-app.use(express.favicon());
+app.use(favicon());
 
 // custom log format
 if ('test' != process.env.NODE_ENV)
-  app.use(express.logger(':method :url'));
+  app.use(logger(':method :url'));
 
 // parses request cookies, populating
 // req.cookies and req.signedCookies
-// when the secret is passed, used 
+// when the secret is passed, used
 // for signing the cookies.
-app.use(express.cookieParser('my secret here'));
+app.use(cookieParser('my secret here'));
 
 // parses json, x-www-form-urlencoded, and multipart/form-data
-app.use(express.bodyParser());
+app.use(bodyParser());
 
 app.get('/', function(req, res){
   if (req.cookies.remember) {