+++ test/req.signedCookies.js
@@ -1,13 +1,14 @@
 
 var express = require('../')
-  , request = require('./support/http');
+  , request = require('./support/http')
+  , cookieParser = require('cookie-parser')
 
 describe('req', function(){
   describe('.signedCookies', function(){
     it('should return a signed JSON cookie', function(done){
       var app = express();
 
-      app.use(express.cookieParser('secret'));
+      app.use(cookieParser('secret'));
 
       app.use(function(req, res){
         if ('/set' == req.path) {